Move collection to resources folder
