+++ omise/test/test_charge.py
@@ -19,7 +19,7 @@ class ChargeTest(_ResourceMixin, unittest.TestCase):
         return Source
 
     def _getCollectionClass(self):
-        from .. import Collection
+        from omise import Collection
         return Collection
 
     def _makeOne(self):
