+++ omise/test/test_receipt.py
@@ -11,7 +11,7 @@ class ReceiptTest(_ResourceMixin, unittest.TestCase):
         return Receipt
 
     def _getCollectionClass(self):
-        from .. import Collection
+        from omise import Collection
         return Collection
 
     def _makeOne(self):
