+++ omise/test/test_schedule.py
@@ -11,7 +11,7 @@ class ScheduleTest(_ResourceMixin, unittest.TestCase):
         return Schedule
 
     def _getCollectionClass(self):
-        from .. import Collection
+        from omise import Collection
         return Collection
 
     def _makeOne(self):
