+++ omise/test/test_transaction.py
@@ -11,7 +11,7 @@ class TransactionTest(_ResourceMixin, unittest.TestCase):
         return Transaction
 
     def _getCollectionClass(self):
-        from .. import Collection
+        from omise import Collection
         return Collection
 
     @mock.patch('requests.get')
